// tests/types/user.types.ts
export type UserData = {
  name: string;
  email: string;
  password: string;
  title: string;
  firstName: string;
  lastName: string;
  birthday: string;
  birthmonth: string;
  birthyear: string;
  company: string;
  address: string;
  city: string;
  state: string;
  country: string;
  zipcode: string;
  mobileNumber: string;
};